expand the game state update to include opponent data
I would like to augment the capablities of the bot players, this first
requires that the bots get access to more data about the game state.

to start, let's update the bot with the opponents played cards, and the
opponents tableu data.

i believe that after each opponent's hand has been seen once, we should
be able to track that hand as it moves across the table. however, i
believe it will be best to wait until later to add that functionality in
the name of a simpler preliminary change.